requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6
openpyxl>=3.1.0
requests>=2.31.0
pyyaml>=6.0.1
pytest>=7.4.0
